betaw3/requirements.txt
Flask
web3
python-dotenv
face_recognition
sqlalchemy
solcx
